Finished removing math.ts
