--- packages/excalidraw/math.ts
@@ -1,205 +1,0 @@
-import type { NormalizedZoomValue, NullableGridSize, Zoom } from "./types";
-import {
-  DEFAULT_ADAPTIVE_RADIUS,
-  LINE_CONFIRM_THRESHOLD,
-  DEFAULT_PROPORTIONAL_RADIUS,
-  ROUNDNESS,
-} from "./constants";
-import type {
-  ExcalidrawElement,
-  ExcalidrawLinearElement,
-} from "./element/types";
-import type { Bounds } from "./element/bounds";
-import {
-  type LocalPoint,
-  type GlobalPoint,
-  point,
-  pointDistance,
-} from "../math";
-
-export const adjustXYWithRotation = (
-  sides: {
-    n?: boolean;
-    e?: boolean;
-    s?: boolean;
-    w?: boolean;
-  },
-  x: number,
-  y: number,
-  angle: number,
-  deltaX1: number,
-  deltaY1: number,
-  deltaX2: number,
-  deltaY2: number,
-): [number, number] => {
-  const cos = Math.cos(angle);
-  const sin = Math.sin(angle);
-  if (sides.e && sides.w) {
-    x += deltaX1 + deltaX2;
-  } else if (sides.e) {
-    x += deltaX1 * (1 + cos);
-    y += deltaX1 * sin;
-    x += deltaX2 * (1 - cos);
-    y += deltaX2 * -sin;
-  } else if (sides.w) {
-    x += deltaX1 * (1 - cos);
-    y += deltaX1 * -sin;
-    x += deltaX2 * (1 + cos);
-    y += deltaX2 * sin;
-  }
-
-  if (sides.n && sides.s) {
-    y += deltaY1 + deltaY2;
-  } else if (sides.n) {
-    x += deltaY1 * sin;
-    y += deltaY1 * (1 - cos);
-    x += deltaY2 * -sin;
-    y += deltaY2 * (1 + cos);
-  } else if (sides.s) {
-    x += deltaY1 * -sin;
-    y += deltaY1 * (1 + cos);
-    x += deltaY2 * sin;
-    y += deltaY2 * (1 - cos);
-  }
-  return [x, y];
-};
-
-export const getPointOnAPath = <P extends LocalPoint | GlobalPoint>(
-  point: P,
-  path: P[],
-) => {
-  const [px, py] = point;
-  const [start, ...other] = path;
-  let [lastX, lastY] = start;
-  let kLine: number = 0;
-  let idx: number = 0;
-
-  // if any item in the array is true, it means that a point is
-  // on some segment of a line based path
-  const retVal = other.some(([x2, y2], i) => {
-    // we always take a line when dealing with line segments
-    const x1 = lastX;
-    const y1 = lastY;
-
-    lastX = x2;
-    lastY = y2;
-
-    // if a point is not within the domain of the line segment
-    // it is not on the line segment
-    if (px < x1 || px > x2) {
-      return false;
-    }
-
-    // check if all points lie on the same line
-    // y1 = kx1 + b, y2 = kx2 + b
-    // y2 - y1 = k(x2 - x2) -> k = (y2 - y1) / (x2 - x1)
-
-    // coefficient for the line (p0, p1)
-    const kL = (y2 - y1) / (x2 - x1);
-
-    // coefficient for the line segment (p0, point)
-    const kP1 = (py - y1) / (px - x1);
-
-    // coefficient for the line segment (point, p1)
-    const kP2 = (py - y2) / (px - x2);
-
-    // because we are basing both lines from the same starting point
-    // the only option for collinearity is having same coefficients
-
-    // using it for floating point comparisons
-    const epsilon = 0.3;
-
-    // if coefficient is more than an arbitrary epsilon,
-    // these lines are nor collinear
-    if (Math.abs(kP1 - kL) > epsilon && Math.abs(kP2 - kL) > epsilon) {
-      return false;
-    }
-
-    // store the coefficient because we are goint to need it
-    kLine = kL;
-    idx = i;
-
-    return true;
-  });
-
-  // Return a coordinate that is always on the line segment
-  if (retVal === true) {
-    return { x: point[0], y: kLine * point[0], segment: idx };
-  }
-
-  return null;
-};
-
-// Checks if the first and last point are close enough
-// to be considered a loop
-export const isPathALoop = (
-  points: ExcalidrawLinearElement["points"],
-  /** supply if you want the loop detection to account for current zoom */
-  zoomValue: Zoom["value"] = 1 as NormalizedZoomValue,
-): boolean => {
-  if (points.length >= 3) {
-    const [first, last] = [points[0], points[points.length - 1]];
-    const distance = pointDistance(first, last);
-
-    // Adjusting LINE_CONFIRM_THRESHOLD to current zoom so that when zoomed in
-    // really close we make the threshold smaller, and vice versa.
-    return distance <= LINE_CONFIRM_THRESHOLD / zoomValue;
-  }
-  return false;
-};
-
-// Returns whether `q` lies inside the segment/rectangle defined by `p` and `r`.
-// This is an approximation to "does `q` lie on a segment `pr`" check.
-export const isPointWithinBounds = <P extends GlobalPoint | LocalPoint>(
-  p: P,
-  q: P,
-  r: P,
-) => {
-  return (
-    q[0] <= Math.max(p[0], r[0]) &&
-    q[0] >= Math.min(p[0], r[0]) &&
-    q[1] <= Math.max(p[1], r[1]) &&
-    q[1] >= Math.min(p[1], r[1])
-  );
-};
-
-// TODO: Rounding this point causes some shake when free drawing
-export const getGridPoint = (
-  x: number,
-  y: number,
-  gridSize: NullableGridSize,
-): [number, number] => {
-  if (gridSize) {
-    return [
-      Math.round(x / gridSize) * gridSize,
-      Math.round(y / gridSize) * gridSize,
-    ];
-  }
-  return [x, y];
-};
-
-export const getCornerRadius = (x: number, element: ExcalidrawElement) => {
-  if (
-    element.roundness?.type === ROUNDNESS.PROPORTIONAL_RADIUS ||
-    element.roundness?.type === ROUNDNESS.LEGACY
-  ) {
-    return x * DEFAULT_PROPORTIONAL_RADIUS;
-  }
-
-  if (element.roundness?.type === ROUNDNESS.ADAPTIVE_RADIUS) {
-    const fixedRadiusSize = element.roundness?.value ?? DEFAULT_ADAPTIVE_RADIUS;
-
-    const CUTOFF_SIZE = fixedRadiusSize / DEFAULT_PROPORTIONAL_RADIUS;
-
-    if (x <= CUTOFF_SIZE) {
-      return x * DEFAULT_PROPORTIONAL_RADIUS;
-    }
-
-    return fixedRadiusSize;
-  }
-
-  return 0;
-};
-
-export const getCenterForElement = (element: ExcalidrawElement): GlobalPoint =>
-  point(element.x + element.width / 2, element.y + element.height / 2);
