+++ packages/math/point.ts
@@ -209,3 +209,18 @@ export const pointScaleFromOrigin = <P extends GlobalPoint | LocalPoint>(
   mid: P,
   multiplier: number,
 ) => pointTranslate(mid, vectorScale(vectorFromPoint(p, mid), multiplier));
+
+// Returns whether `q` lies inside the segment/rectangle defined by `p` and `r`.
+// This is an approximation to "does `q` lie on a segment `pr`" check.
+export const isPointWithinBounds = <P extends GlobalPoint | LocalPoint>(
+  p: P,
+  q: P,
+  r: P,
+) => {
+  return (
+    q[0] <= Math.max(p[0], r[0]) &&
+    q[0] >= Math.min(p[0], r[0]) &&
+    q[1] <= Math.max(p[1], r[1]) &&
+    q[1] >= Math.min(p[1], r[1])
+  );
+};
